Use new routine to store each line into buffer

+++ MHD/C_src/Obj_C/PsfController.m
@@ -690,7 +690,7 @@
 - (IBAction)PsfZeroLineSwitchAction:(id)sender;
 {
     struct kemoviewer_type *kemo_sgl = [_kmv KemoViewPointer];
-    kemoview_set_PSF_draw_flags(ZEROGRID_TOGGLE, self.PSFZerolineSwitch, kemo_sgl);
+    kemoview_set_PSF_draw_flags(ZEROGRID_TOGGLE, (int) self.PSFZerolineSwitch, kemo_sgl);
 	self.PSFLineSwitch = self.PSFZerolineSwitch + self.PSFIsolineSwitch;
     [self UpdateCurrentPsfMenu:kemo_sgl];
 	[_metalView UpdateImage:kemo_sgl];
@@ -699,7 +699,7 @@
 - (IBAction)PsfColorbarSwitchAction:(id)sender;
 {
     struct kemoviewer_type *kemo_sgl = [_kmv KemoViewPointer];
-    kemoview_set_PSF_draw_flags(COLORBAR_TOGGLE, self.PSFColorbarSwitch, kemo_sgl);
+    kemoview_set_PSF_draw_flags(COLORBAR_TOGGLE, (int) self.PSFColorbarSwitch, kemo_sgl);
 	[_metalView UpdateImage:kemo_sgl];
 }
